pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "nano-biomni"
version = "0.1.0"
description = "Standalone biomedical research agent — adapts Biomni's ReAct loop for paper replication"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3",
    "langchain-core>=0.3",
    "langgraph>=0.2",
    "langchain-ollama>=0.2",
    "langchain-openai>=0.2",
    "langchain-anthropic>=0.3",
    "PyMuPDF>=1.23",
    "PyPDF2>=3.0",
    "arxiv>=2.0",
    "biopython>=1.83",
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "googlesearch-python>=1.2",
    "python-dotenv>=1.0",
    "pydantic>=2.6",
]

[project.scripts]
nano-biomni = "nano_biomni.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["nano_biomni*"]
